+++ test/integration/tests/attestation.sh
@@ -92,7 +92,7 @@ tpm2_pcrreset -Q $debug_pcr
 tpm2_pcrextend -Q $debug_pcr:sha256=$rand_pcr_value
 tpm2_pcrlist -Q
 getrandom 20
-tpm2_quote -Q -C $context_ak -L $digestAlg:$debug_pcr_list -q $loaded_randomness -m $output_quote -s $output_quotesig -p $output_quotepcr -g $digestAlg -P "$akpw"
+tpm2_quote -Q -C $context_ak -l $digestAlg:$debug_pcr_list -q $loaded_randomness -m $output_quote -s $output_quotesig -f $output_quotepcr -g $digestAlg -P "$akpw"
 
 
 # Verify quote